c++, contracts: Use the saved_scope to carry state not globals.
This uses state local to the contract scopes to carry the information
on const-ification and whether the current contract is a post.

Note that if, for example, we had a lambda with a precondition in the
defintion of a post-condition, the use of increment/decrement of a
"postcondition" variable is not going to work properly.

We also use a contract scope during template substitution so that we
can check for that scope when deciding how to process.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>