c++, contracts: Handle contract-specific location wrappers.
Contract conditions use location wrappers on more tree types than
the generic code.  This means that we need to take care of them
when tsubst-ing, otherwise the geneic code is getting given un-
expected input (and, predicatably, it does not work).

This is a more general problem upstream - but "just applying
wrappers" to more trees in the generic code is going to need
a lot of work (it fails quite spectacularly) unrelated to
contracts.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>