c++, contracts: Rework  __builtin_contract_violation_type.
This does all the work needed to have the builtin type but does not
(yet) publish it - since we are still figuring out the details of
the ABI - however,

In this version, we no longer try to declare or build a
std::source_location since this creates quite some difficulty with
needing the declaration whether <source_location> is included or not.

Instead, we save a type-erased pointer to the implementation data.
Since clang and GCC actually use the same layout here, there is a move
to add this to the Itanium spec at that point we will be able to use
some __cxa_xxxxx_t * as the pointer type.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>