c++, contracts: Factor contract_assert code.
This puts the contract_assert parse code together with the other p2900
parsers.  It is also the first step in removing the attribute wrappers
from p2900 cases (since the contract_assert is a regular statement we
can just handle it as such and have no need to generate an attribute
and subsequently expand it.

Small adjustments to error-recovery and using a standard check for the
closing semicolon needs adjustment of expected test-case output.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>