c++, contracts: Remove function contract specifiers from attributes.
They are not attributes, and treating them as such is hindering progress.

In the revised design, they are kept as on-the-side data for function
decls that have contract specifiers (in a similar way to other out-of-
band specifiers).

Since we are currently still supporting attribute-based contracts for
the cxx2a case, and there is some share code, in the first instance we
have a number of cases wherewe have to select the source of the data
depending on flag_contracts_nonattr.  Eventually, these cases will be
removed (1) because we want to do the non-attribute work a different
way and (2) because we will remove the support for the attribute case
which was not adopted in the WD.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>