c++, contracts: Separation of contracts implementations.
We have, at present, a mixture of pieces of declatations in cp-tree.h
and contracts.h (which is also included in cp-tree.h).  Remove all the
contracts information from cp-tree.h and re-organise it in contracts.h.

Then include contracts.h where it is needed only.

In a similar manner we have two implementations in contracts.cc where
one of them (cxx2a contracts) is intended to be deprecated and finally
removed.  However there is currently shared support code and common
public APIs.  This reorganises the code with the following head comment.

"
To facilitate maintenance, please keep the general ordering of content:

 shared local support functions
  cxx2a implementation:
    local support functions
    cxx2a-only public API
  cxx26 implementation:
    local support functions
    cxx26-only public API
 shared public API.
"

In the process, we identified some APIs that are actually local to
contracts.cc and these have been made static functions and removed
from the public API.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>