c++, contracts: Avoid unnecessary specifier map entries.
When we have a specialisation, it might be given contract specifiers
(depending on whether they are present on the template, as a side-effect
of the instantiation process).  However, these added specifiers are then
deleted - since the presence of (and spelling of) specifiers on
specialisations is independent of the original template.

We were doing this by adding empty contracts to the specialisation.
This has an unfortunate side-effect of adding decls to the specifier
lookup map, even if the decl had no contracts already.  This can play
badly with GC and hash collisions.

Solve this by adding an interface that only deletes specifiers on
functions already registered.

This is subtely different from the case when we are about to
ggc_delete a decl (in that case we actually also want to remove the
decl from the map too).  So renamed that interface to say that it
will remove the decl.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>